Move ``SetupTeardownContext`` to Task SDK
Part of https://github.com/apache/airflow/issues/51671

The code is only used in Task SDK anyway fro setup and teardown decorators
